NAS-3716 - Generate new client for versioned api
 - move all calls to api/v1/profile into factory

JIRA: FET-1033
