+++ libs/api-client-tiger/src/client.ts
@@ -22,6 +22,7 @@ import {
     MetadataRequestArgs,
     tigerEntitiesObjectsClientFactory,
 } from "./entitiesObjects";
+import { tigerProfileClientFactory, IUserProfile, ProfileApiInterface } from "./profile";
 
 export {
     tigerExecutionClientFactory,
@@ -31,6 +32,7 @@ export {
     tigerValidObjectsClientFactory,
     tigerLayoutClientFactory,
     tigerAfmExplainClientFactory,
+    tigerProfileClientFactory,
     MetadataConfiguration,
     MetadataConfigurationParameters,
     MetadataBaseApi,
@@ -39,6 +41,8 @@ export {
     LabelElementsConfigurationParameters,
     LabelElementsBaseApi,
     LabelElementsRequestArgs,
+    ProfileApiInterface,
+    IUserProfile,
 };
 
 export interface ITigerClient {
@@ -50,6 +54,7 @@ export interface ITigerClient {
     explain: ReturnType<typeof tigerAfmExplainClientFactory>;
     declarativeLayout: ReturnType<typeof tigerLayoutClientFactory>;
     entities: ReturnType<typeof tigerEntitiesObjectsClientFactory>;
+    profile: ReturnType<typeof tigerProfileClientFactory>;
 
     /**
      * Updates tiger client to send the provided API TOKEN in `Authorization` header of all
@@ -73,6 +78,7 @@ export const tigerClientFactory = (axios: AxiosInstance): ITigerClient => {
     const declarativeLayout = tigerLayoutClientFactory(axios);
     const explain = tigerAfmExplainClientFactory(axios);
     const entities = tigerEntitiesObjectsClientFactory(axios);
+    const profile = tigerProfileClientFactory(axios);
 
     return {
         axios,
@@ -83,6 +89,7 @@ export const tigerClientFactory = (axios: AxiosInstance): ITigerClient => {
         declarativeLayout,
         explain,
         entities,
+        profile,
         setApiToken: (token: string | undefined): void => {
             setAxiosAuthorizationToken(axios, token);
         },
