+++ libs/sdk-backend-tiger/src/backend/user/index.ts
@@ -2,7 +2,7 @@
 import { IUser, IUserService, IUserSettingsService } from "@gooddata/sdk-backend-spi";
 import { TigerUserSettingsService } from "./settings";
 import { TigerAuthenticatedCallGuard } from "../../types";
-import { convertUser, IUserProfile } from "../../convertors/fromBackend/UsersConverter";
+import { convertUser } from "../../convertors/fromBackend/UsersConverter";
 
 export class TigerUserService implements IUserService {
     constructor(private readonly authCall: TigerAuthenticatedCallGuard) {}
@@ -12,10 +12,8 @@ export class TigerUserService implements IUserService {
     }
 
     public async getUser(): Promise<IUser> {
-        // TODO: replace with direct call of TigerClient (once methods are generated from OpenAPI)
         return this.authCall(async (client) => {
-            const profile = await client.axios.get<IUserProfile>("/api/v1/profile");
-            return convertUser(profile.data);
+            return convertUser(await client.profile.getCurrent());
         });
     }
 }
