+++ libs/sdk-backend-tiger/src/convertors/fromBackend/UsersConverter.ts
@@ -1,15 +1,6 @@
-// (C) 2019-2021 GoodData Corporation
+// (C) 2019-2022 GoodData Corporation
 import { IUser, uriRef } from "@gooddata/sdk-model";
-
-export interface IUserProfile {
-    name: string;
-    userId: string;
-    organizationName: string;
-    links: {
-        user: string;
-        organization: string;
-    };
-}
+import { IUserProfile } from "@gooddata/api-client-tiger";
 
 /**
  * To perserve the typing and bootstrap concept, we are using firstName
